Created routes and moved postData to controllers
