+++ routes/match.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router();
+const matchController = require("../controllers/match");
+
+router.post("", matchController.createMatch);
+
+module.exports = router;
